Added support for showing diff with tree of tables, where tables can be made hidden/shown

+++ packages/vuerd/src/core/parser/ParserToJson.ts
@@ -67,25 +67,34 @@ function reshape(
     switch (statement.type) {
       case 'create.table':
         const table = statement;
-        if (table.name) {
+        const duplicateTable = findByName(shape.tables, table.name);
+        if (!duplicateTable && table.name) {
           shape.tables.push(table);
         }
         break;
       case 'create.index':
         const index = statement;
-        if (index.tableName && index.columns.length) {
+        const duplicateIndex = findByName(shape.indexes, index.name);
+        if (!duplicateIndex && index.tableName && index.columns.length) {
           shape.indexes.push(index);
         }
         break;
       case 'alter.table.add.primaryKey':
         const primaryKey = statement;
-        if (primaryKey.name && primaryKey.columnNames.length) {
+        const duplicatePK = findByName(shape.primaryKeys, primaryKey.name);
+        if (!duplicatePK && primaryKey.name && primaryKey.columnNames.length) {
           shape.primaryKeys.push(primaryKey);
         }
         break;
       case 'alter.table.add.foreignKey':
         const foreignKey = statement;
+        const duplicateFK = findByConstraintName(
+          shape.foreignKeys,
+          foreignKey.constraintName
+        );
+
         if (
+          !duplicateFK &&
           foreignKey.name &&
           foreignKey.columnNames.length &&
           foreignKey.refTableName &&
@@ -97,31 +106,57 @@ function reshape(
         break;
       case 'alter.table.add.unique':
         const unique = statement;
-        if (unique.name && unique.columnNames.length) {
+        const duplicateUnique = findByName(shape.uniques, unique.name);
+        if (!duplicateUnique && unique.name && unique.columnNames.length) {
           shape.uniques.push(unique);
         }
         break;
       case 'alter.table.add.column':
         const addColumns = statement;
-        if (addColumns.name && addColumns.columns.length) {
+        const duplicateAddColumns = findByName(
+          shape.addColumns,
+          addColumns.name
+        );
+        if (
+          !duplicateAddColumns &&
+          addColumns.name &&
+          addColumns.columns.length
+        ) {
           shape.addColumns.push(addColumns);
         }
         break;
       case 'alter.table.drop.column':
         const dropColumns = statement;
-        if (dropColumns.name && dropColumns.columns.length) {
+        const duplicateDropColumns = findByName(
+          shape.dropColumns,
+          dropColumns.name
+        );
+        if (
+          !duplicateDropColumns &&
+          dropColumns.name &&
+          dropColumns.columns.length
+        ) {
           shape.dropColumns.push(dropColumns);
         }
         break;
       case 'drop.table':
         const dropTable = statement;
-        if (dropTable.name) {
+        const duplicateDropTable = findByName(shape.dropTable, dropTable.name);
+        if (!duplicateDropTable && dropTable.name) {
           shape.dropTable.push(dropTable);
         }
         break;
       case 'alter.table.drop.foreignKey':
         const dropForeignKey = statement;
-        if (dropForeignKey.name && dropForeignKey.baseTableName) {
+        const duplicateDropFK = findByName(
+          shape.dropForeignKeys,
+          dropForeignKey.name
+        );
+        if (
+          !duplicateDropFK &&
+          dropForeignKey.name &&
+          dropForeignKey.baseTableName
+        ) {
           shape.dropForeignKeys.push(dropForeignKey);
         }
         break;
@@ -137,6 +172,18 @@ function findByName<T extends { name: string }>(
 ): T | null {
   for (const item of list) {
     if (item.name.toUpperCase() === name.toUpperCase()) {
+      return item;
+    }
+  }
+  return null;
+}
+
+function findByConstraintName<T extends { constraintName: string }>(
+  list: T[],
+  constraintName: string
+): T | null {
+  for (const item of list) {
+    if (item.constraintName.toUpperCase() === constraintName.toUpperCase()) {
       return item;
     }
   }
@@ -201,6 +248,7 @@ function mergeTable(shape: Shape): CreateTable[] {
         refTableName: foreignKey.refTableName,
         refColumnNames: foreignKey.refColumnNames,
         constraintName: foreignKey.constraintName,
+        visible: foreignKey.visible,
       });
     }
   });
@@ -209,7 +257,10 @@ function mergeTable(shape: Shape): CreateTable[] {
     const table = findByName(tables, addColumn.name);
     if (table) {
       addColumn.columns.forEach(column => {
-        table.columns.push(column);
+        const duplicateColumn = findByName(table.columns, column.name);
+        if (!duplicateColumn) {
+          table.columns.push(column);
+        }
       });
     }
   });
@@ -229,8 +280,6 @@ function mergeTable(shape: Shape): CreateTable[] {
     tables = tables.filter(table => table.name !== dropTable.name);
   });
 
-  console.log(dropForeignKeys);
-
   dropForeignKeys.forEach(dropForeignKey => {
     const table = findByName(tables, dropForeignKey.baseTableName);
     if (table) {
@@ -248,12 +297,7 @@ function mergeTable(shape: Shape): CreateTable[] {
  * @param snaphot Latest snapshot
  * @returns Shape with all statements needed to replicate latest snapshot
  */
-function snapshotToShape({
-  canvas,
-  memo,
-  table,
-  relationship,
-}: ExportedStore): Shape {
+function snapshotToShape({ table, relationship }: ExportedStore): Shape {
   const shape: Shape = {
     tables: [],
     indexes: [],
@@ -287,6 +331,7 @@ function snapshotToShape({
         foreignKeys: [],
         indexes: [],
         name: table.name,
+        visible: table.visible,
       };
 
       return createTable;
@@ -337,6 +382,7 @@ function snapshotToShape({
         refTableName: refTable?.name || '',
         refColumnNames: refColumnNames,
         constraintName: relationship.constraintName,
+        visible: relationship.visible,
       };
 
       return fk;
@@ -432,6 +478,7 @@ function createTable(
           widthComment: SIZE_MIN_WIDTH,
           zIndex: 2,
         },
+    visible: table.visible === undefined ? true : table.visible,
   } as any;
 
   const widthName = helper.getTextWidth(newTable.name);
@@ -532,6 +579,12 @@ function createRelationship(data: ExportedStore, tables: CreateTable[]) {
               }
             });
 
+            if (startTable.visible && endTable.visible) {
+              foreignKey.visible = true;
+            } else {
+              foreignKey.visible = false;
+            }
+
             data.relationship.relationships.push({
               id: uuid(),
               identification: !endColumns.some(column => !column.ui.pfk),
@@ -551,6 +604,7 @@ function createRelationship(data: ExportedStore, tables: CreateTable[]) {
                 direction: 'top',
               },
               constraintName: foreignKey.constraintName,
+              visible: foreignKey.visible,
             });
           }
         });
